Included Final Writeup on Confusion Feature
This is the final writeup for this project, which was the culmination of months of work on my end. While I would've liked to expand the expriment further, I am happy with where this writeup stands as a strong proof of concept for Confusion. 

In short, confusion is an anomaly detection mechanism focused on latent neural network activation layers instead of the input layer.  I used a feed-forward autoencoder for this task , but there are other ways I explored in the writeup.